+++ pyriemann/artifact_detection.py
